+++ documentation/src/icons/components/Sequences.tsx
@@ -1,0 +1,8 @@
+/// <Reference path="../icon.d.ts" />
+import type { h as JSXFactory } from '@stencil/core';
+
+export const Sequences = (h: typeof JSXFactory) => (props: any) => (
+    <svg {...props} viewBox="0 0 24 24" aria-hidden="true">
+        <path d="M3 17h18v-2H3v2m0 3h18v-1H3v1m0-7h18v-3H3v3m0-9v4h18V4H3Z" />
+    </svg>
+);