+++ packages/sequences/src/components.d.ts
@@ -16,11 +16,6 @@ export namespace Components {
          */
         "number"?: number;
     }
-    /**
-     * Sequence
-     */
-    interface SequenceDemo {
-    }
 }
 declare global {
     /**
@@ -33,18 +28,8 @@ declare global {
         prototype: HTMLKurrentSequenceElement;
         new (): HTMLKurrentSequenceElement;
     };
-    /**
-     * Sequence
-     */
-    interface HTMLSequenceDemoElement extends Components.SequenceDemo, HTMLStencilElement {
-    }
-    var HTMLSequenceDemoElement: {
-        prototype: HTMLSequenceDemoElement;
-        new (): HTMLSequenceDemoElement;
-    };
     interface HTMLElementTagNameMap {
         "kurrent-sequence": HTMLKurrentSequenceElement;
-        "sequence-demo": HTMLSequenceDemoElement;
     }
 }
 declare namespace LocalJSX {
@@ -58,14 +43,8 @@ declare namespace LocalJSX {
          */
         "number"?: number;
     }
-    /**
-     * Sequence
-     */
-    interface SequenceDemo {
-    }
     interface IntrinsicElements {
         "kurrent-sequence": KurrentSequence;
-        "sequence-demo": SequenceDemo;
     }
 }
 export { LocalJSX as JSX };
@@ -77,10 +56,6 @@ declare module "@stencil/core" {
              * If number isn't set, it will display one at random.
              */
             "kurrent-sequence": LocalJSX.KurrentSequence & JSXBase.HTMLAttributes<HTMLKurrentSequenceElement>;
-            /**
-             * Sequence
-             */
-            "sequence-demo": LocalJSX.SequenceDemo & JSXBase.HTMLAttributes<HTMLSequenceDemoElement>;
         }
     }
 }